Miglioramenti vari al codice e minor bug fix.

+++ app/Http/Controllers/FixingController.php
@@ -4,17 +4,25 @@ namespace Tentazioninoro\Http\Controllers;
 
 use Auth;
 use Config;
-use DB;
 use Debugbar;
 use \Input as Input;
-use Redirect;
 use View;
 use Illuminate\Http\Request;
 use Tentazioninoro\Customer;
 use Tentazioninoro\Fixing;
 use Tentazioninoro\Jewel;
+use Tentazioninoro\User;
 
 class FixingController extends Controller {
+
+    private function isAllowedAccess() {
+	if (Auth::check() && Auth::user()->permissions === \Config::get('constants.permission.FIXINGS')) {
+	    return true;
+	} else {
+	    redirect('/access-not-allowed');
+	    return false;
+	}
+    }
 
     /**
      * Display a listing of the resource.
@@ -22,10 +30,12 @@ class FixingController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
-	$userId = Auth::id();
-	$fixingList = Fixing::where('user_id', $userId)->get();
-	Debugbar::info($fixingList);
-	return View::make('fixing/index')->with('fixingList', $fixingList);
+	if ($this->isAllowedAccess()) {
+	    $userId = Auth::id();
+	    $fixingList = Fixing::where('user_id', $userId)->get();
+	    Debugbar::info($fixingList);
+	    return View::make('fixing/index')->with('fixingList', $fixingList);
+	}
     }
 
     /**
@@ -34,10 +44,43 @@ class FixingController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function create() {
-	$user = Auth::user();
-	$fixingList = Fixing::where('user_id', $user->id)->orderBy('customer_id', 'asc')->get();
-	Debugbar::info($fixingList);
-	return View::make('fixing/create', ['user' => $user]);
+	if ($this->isAllowedAccess()) {
+//	    $user = Auth::user();
+//	    $fixingList = Fixing::where('user_id', $user->id)->orderBy('customer_id', 'asc')->get();
+//	    Debugbar::info($fixingList);
+//	    return View::make('fixing/create', ['user' => $user]);
+
+	    $userId = Auth::id();
+	    $customersIds = User::find($userId)->customers()->get(); //->groupBy('customer_id');
+	    Debugbar::info($customersIds);
+	    $customerList = array();
+	    foreach ($customersIds as $customerId) {
+		Debugbar::info("customerId - start", $customerId, "customerId - end");
+		$identityDocument = Customer::find($customerId->customer_id)->identityDocument;
+		$customerList[$customerId->customer_id] = $identityDocument->name . " " . $identityDocument->surname;
+		//            $customerList[] = $identityDocument;
+	    }
+	    //        $customerList = [1,2,43];
+	    //            Debugbar::info("fixingList - start", $fixingList, "fixingList - end");
+	    //            Debugbar::info($customerList);
+	    if (isset($customerId)) {
+		$customer = Customer::where('id', $customerId)->get();
+	    } else {
+		$customer = new Customer;
+	    }
+
+	    $fixing = new Fixing;
+	    //        Debugbar::info($user);
+	    //        Debugbar::info($fixing);
+
+	    $data = array(
+		'showCustomerList' => true,
+		'customerList' => $customerList,
+		'fixing' => $fixing,
+		'customer' => $customer,
+	    );
+	    return View::make('fixing/create')->with('data', $data);
+	}
     }
 
     /**
@@ -47,43 +90,45 @@ class FixingController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request) {
-	$this->validate($request, Fixing::$rules);
+	if ($this->isAllowedAccess()) {
+	    $this->validate($request, Fixing::$rules);
+	    $fixing = $request->all();
+	    $id = Auth::id();
+	    if (Input::hasFile('path_photo')) {
+//		$uploadDirectory = "uploads";
+//		$file = Input::file('path_photo');
+//		$file->move($uploadDirectory, $file->getClientOriginalName());
+//		$path_to_photo = $uploadDirectory . "\\" . $file->getClientOriginalName();
+		$path = $request->file('path_photo')->store('fixings');
+		$jewelData = array(
+		    "typology" => $fixing["typology"],
+		    "weight" => $fixing["weight"],
+		    "metal" => $fixing["metal"],
+		    "path_photo" => $path,
+		);
+		$jewel = Jewel::create($jewelData);
+		//            JewelController::save_photo_into_cloud($path_to_photo);
 
-	$fixing = $request->all();
-	$id = Auth::id();
-	if (Input::hasFile('path_photo')) {
-	    $uploadDirectory = "uploads";
-	    $file = Input::file('path_photo');
-	    $file->move($uploadDirectory, $file->getClientOriginalName());
-            $path_to_photo = $uploadDirectory . "\\" . $file->getClientOriginalName();
-	    $jewelData = array(
-		"typology" => $fixing["typology"],
-		"weight" => $fixing["weight"],
-		"metal" => $fixing["metal"],
-		"path_photo" => $path_to_photo,
-	    );
-	    $jewel = Jewel::create($jewelData);
-            JewelController::save_photo_into_cloud($path_to_photo);
+		$fixingData = array(
+		    "user_id" => $id,
+		    "customer_id" => $fixing["customer_id"],
+		    "jewel_id" => $jewel->id,
+		    "description" => $fixing["description"],
+		    "deposit" => $fixing["deposit"],
+		    "estimate" => $fixing["estimate"],
+		    "notes" => $fixing["notes"],
+		    "state" => Config::get('constants.fixing.state.NOT_YET_STARTED'),
+		);
 
-	    $fixingData = array(
-		"user_id" => $id,
-		"customer_id" => $fixing["customer_id"],
-		"jewel_id" => $jewel->id,
-		"description" => $fixing["description"],
-		"deposit" => $fixing["deposit"],
-		"estimate" => $fixing["estimate"],
-		"notes" => $fixing["notes"],
-		"state" => Config::get('constants.fixing.state.not_yet_started'),
-	    );
+		//        var_dump($jewel);
+		//        echo "<br>";
+		//        echo $jewel->id;
+		//        echo "<br>";
+		Fixing::create($fixingData);
+	    }
 
-    //        var_dump($jewel);
-    //        echo "<br>";
-    //        echo $jewel->id;
-    //        echo "<br>";
-	    Fixing::create($fixingData);
+	    return redirect(route('fixing.index'));
 	}
-	
-	return Redirect::to(route('fixing.index'));
     }
 
     /**
@@ -92,13 +137,30 @@ class FixingController extends Controller {
      * @param  \Tentazioninoro\Fixing  $fixing
      * @return \Illuminate\Http\Response
      */
-    public function show() {
-	$userId = Auth::id();
-//	$fixingList = Fixing::where('user_id', $user->id)->orderBy('customer_id', 'asc')->get();
-//	Debugbar::info($fixingList);
-	$fixingList = Fixing::where('user_id', $userId)->get();
-	Debugbar::info($userId);
-	return View::make('fixing/index')->with('fixingList', $fixingList);
+    public function show($fixingId) {
+	if ($this->isAllowedAccess()) {
+	    $userId = Auth::id();
+	    $customersIds = User::find($userId)->customers()->get(); //->groupBy('customer_id');
+
+	    $customer = new Customer;
+	    $jewel = new Jewel;
+	    $fixing = Fixing::where('id', $fixingId)->get()->first();
+	    $jewel = Jewel::where('id', $fixing->jewel_id)->get()->first();
+//            $customer = Customer::where('id', $fixing->customer_id)->get()->first();
+	    $identityDocument = Customer::find($fixing->customer_id)->identityDocument;
+
+	    $data = array(
+		'showCustomerList' => false,
+		'fixing' => $fixing,
+//                'customer' => $customer,
+		'identityDocument' => $identityDocument,
+		'jewel' => $jewel,
+	    );
+
+	    return View::make('fixing/create')->with('data', $data);
+	} else {
+	    return redirect(route('home'));
+	}
     }
 
     /**
@@ -108,22 +170,24 @@ class FixingController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function showList($customerId = NULL) {
-	$user = Auth::user();
-	$userId = Auth::id();
-
-	if (isset($customerId)) {
-	    $whereArray = [
-		'user_id' => $userId,
-		'customer_id' => $customerId,
-	    ];
+	if ($this->isAllowedAccess()) {
+	    $userId = Auth::id();
+	    if (isset($customerId)) {
+		$whereArray = [
+		    'user_id' => $userId,
+		    'customer_id' => $customerId,
+		];
+	    } else {
+		$whereArray = [
+		    'user_id' => $userId,
+		];
+	    }
+	    $fixingList = Fixing::where($whereArray)->get();
+	    DebugBar::info($fixingList);
+	    return View::make('fixing/index')->with('fixingList', $fixingList);
 	} else {
-	    $whereArray = [
-		'user_id' => $userId,
-	    ];
+//	    return redirect('/access-not-allowed');
 	}
-	$fixingList = Fixing::where($whereArray)->get();
-	DebugBar::info($fixingList);
-	return View::make('fixing/index')->with('fixingList', $fixingList);
     }
 
     /**
@@ -133,7 +197,9 @@ class FixingController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function edit(Fixing $fixing) {
-//	return View::make('fixing/create', ['user' => $user]);
+	if ($this->isAllowedAccess()) {
+//	    return View::make('fixing/create', ['user' => $user]);
+	}
     }
 
     /**
@@ -144,7 +210,9 @@ class FixingController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Fixing $fixing) {
-	//
+	if ($this->isAllowedAccess()) {
+	    
+	}
     }
 
     /**
@@ -155,17 +223,21 @@ class FixingController extends Controller {
      */
     public function destroy($id) {
 //	var_dump($id);
-	Fixing::destroy($id);
-	return Redirect::to(route('fixing.index'));
+	if ($this->isAllowedAccess()) {
+	    Fixing::destroy($id);
+	    return redirect(route('fixing.index'));
+	}
     }
 
     public function destroyFixings(Request $request) {
-	$ids = $request->input('ids');
-        foreach ($ids as $id) {
-            print_r($id);
-            Fixing::destroy($id);
-        }
-	return redirect(route('fixing.index'));
+	if ($this->isAllowedAccess()) {
+	    $ids = $request->input('ids');
+	    foreach ($ids as $id) {
+		print_r($id);
+		Fixing::destroy($id);
+	    }
+	    return redirect(route('fixing.index'));
+	}
     }
 
 }
